+++ _templates/generator/new/prompt.js
@@ -1,0 +1,8 @@
+module.exports = [
+  {
+    type: 'select',
+    name: 'category',
+    message: 'Select hooks category',
+    choices: ['effect', 'state', 'effectstate', 'ref']
+  }
+]