+++ scripts/build.ts
@@ -2,10 +2,10 @@ import type { BuildOptions } from 'esbuild'
 import { build } from 'esbuild'
 import { resolve } from 'path'
 
-import { main, module, peerDependencies } from './package.json'
+import { main, module, peerDependencies } from '../package.json'
 
 const sharedOptions: BuildOptions = {
-  entryPoints: [resolve(__dirname, 'lib', 'index.ts')],
+  entryPoints: [resolve(__dirname, '..', 'lib', 'index.ts')],
   bundle: true,
   minify: true,
   sourcemap: true,