+++ tests/server/useIdleCallbackEffect_test.ts
@@ -1,0 +1,16 @@
+import { renderHook } from '@testing-library/react-hooks/server'
+
+import { useIdleCallbackEffect } from '@/useIdleCallbackEffect'
+
+describe('useIdleCallbackEffect', () => {
+  it('should not throw error', () => {
+    const callback = jest.fn()
+    const { result } = renderHook(() =>
+      useIdleCallbackEffect({
+        callback
+      })
+    )
+
+    expect(result.error).toBeUndefined()
+  })
+})